// nextjs/components/database/schemas/public/InvoiceLineItem.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type {InvoiceId} from './Invoice';

/** Identifier type for public.invoice_line_item */
export type Invoice_line_itemId = string;

/** Represents the table public.invoice_line_item */
export default interface InvoiceLineItem {
    id: Invoice_line_itemId;

    invoiceId: InvoiceId;

    customerId: string;

    company: string;

    invoiceDate: Date;

    campaignName: string;

    tariffId: string;

    description: string;

    quantity: string;

    unitPrice: string;

    lineTotal: string;

    currency: string;
}

/** Represents the initializer for the table public.invoice_line_item */
export interface InvoiceLineItemInitializer {
    /** Default value: uuidv7_sub_ms() */
    id?: Invoice_line_itemId;

    invoiceId: InvoiceId;

    customerId: string;

    company: string;

    invoiceDate: Date;

    campaignName: string;

    tariffId: string;

    description: string;

    quantity: string;

    unitPrice: string;

    lineTotal: string;

    currency: string;
}

/** Represents the mutator for the table public.invoice_line_item */
export interface InvoiceLineItemMutator {
    id?: Invoice_line_itemId;

    invoiceId?: InvoiceId;

    customerId?: string;

    company?: string;

    invoiceDate?: Date;

    campaignName?: string;

    tariffId?: string;

    description?: string;

    quantity?: string;

    unitPrice?: string;

    lineTotal?: string;

    currency?: string;
}